fixed walrus exception handling.

+++ clc/modules/core/src/main/java/edu/ucsb/eucalyptus/cloud/entities/WalrusInfo.java
@@ -20,9 +20,9 @@ public class WalrusInfo {
 	@Column( name = "storage_max_buckets_per_user" )
 	private Integer storageMaxBucketsPerUser;
 	@Column( name = "storage_max_bucket_size_mb" )
-	private Long storageMaxBucketSizeInMB;
+	private Integer storageMaxBucketSizeInMB;
 	@Column( name = "storage_cache_size_mb" )
-	private Long storageMaxCacheSizeInMB;
+	private Integer storageMaxCacheSizeInMB;
 	@Column( name = "storage_snapshot_size_gb" )
 	private Integer storageMaxTotalSnapshotSizeInGb;
 
@@ -31,8 +31,8 @@ public class WalrusInfo {
 	public WalrusInfo(final String name, 
 			final String storageDir,
 			final Integer storageMaxBucketsPerUser,
-			final Long storageMaxBucketSizeInMB,
-			final Long storageMaxCacheSizeInMB,
+			final Integer storageMaxBucketSizeInMB,
+			final Integer storageMaxCacheSizeInMB,
 			final Integer storageMaxTotalSnapshotSizeInGb)
 	{
 		this.name = name;
@@ -71,19 +71,19 @@ public class WalrusInfo {
 		this.storageMaxBucketsPerUser = storageMaxBucketsPerUser;
 	}
 
-	public Long getStorageMaxBucketSizeInMB() {
+	public Integer getStorageMaxBucketSizeInMB() {
 		return storageMaxBucketSizeInMB;
 	}
 
-	public void setStorageMaxBucketSizeInMB( final Long storageMaxBucketSizeInMB ) {
+	public void setStorageMaxBucketSizeInMB( final Integer storageMaxBucketSizeInMB ) {
 		this.storageMaxBucketSizeInMB = storageMaxBucketSizeInMB;
 	}
 
-	public Long getStorageMaxCacheSizeInMB() {
+	public Integer getStorageMaxCacheSizeInMB() {
 		return storageMaxCacheSizeInMB;
 	}
 
-	public void setStorageMaxCacheSizeInMB( final Long storageMaxCacheSizeInMB ) {
+	public void setStorageMaxCacheSizeInMB( final Integer storageMaxCacheSizeInMB ) {
 		this.storageMaxCacheSizeInMB = storageMaxCacheSizeInMB;
 	}
